backend/requirements.txt
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==22.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
build==0.10.0
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
comm==0.2.2
contourpy==1.2.1
crashtest==0.4.1
cryptography==39.0.2
cycler==0.12.1
Cython==0.29.37
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.1
django-cors-headers==4.4.0
django-simple-history==3.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dulwich==0.21.3
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
fonttools==4.53.1
fqdn==1.5.1
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.4
importlib-metadata==6.0.0
installer==0.6.0
ipykernel==6.29.5
ipython==8.26.0
isoduration==20.11.0
jaraco.classes==3.2.3
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keyring==23.13.1
kiwisolver==1.4.5
lockfile==0.12.2
Mako==1.3.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
meson==1.3.2
mistune==3.0.2
more-itertools==9.1.0
msgpack==1.0.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.1
notebook_shim==0.2.4
numpy==2.0.1
overrides==7.7.0
packaging==23.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.8.0
pillow==10.4.0
pipenv==2023.2.18
pkginfo==1.9.6
platformdirs==4.2.2
poetry==1.4.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.21
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.2
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==2.15.1
referencing==0.35.1
requests==2.32.3
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.19.1
scikit-learn==1.5.1
scipy==1.14.1
SecretStorage==3.3.3
Send2Trash==1.8.3
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.1
stack-data==0.6.3
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tomlkit==0.11.6
tornado==6.4.1
traitlets==5.14.3
trove-classifiers==2023.3.9
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==1.26.15
virtualenv==20.26.3
virtualenv-clone==0.5.7
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
zipp==3.15.0
